% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_neptune_endpoint.R
\name{connect_to_neptune_endpoint}
\alias{connect_to_neptune_endpoint}
\title{Title}
\usage{
connect_to_neptune_endpoint(host, port = 8182, https = FALSE,
  type = c("gremlin", "sparql"))
}
\arguments{
\item{port}{Port for the connection (defaults to 8182)}

\item{type}{c('gremlin', 'sparql')}
}
\description{
Title
}
