% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgesdata.R
\docType{data}
\name{Edges}
\alias{Edges}
\title{Edges for the Tinkerpop graph example}
\format{An object of class \code{data.frame} with 6 rows and 6 columns.}
\source{
\href{http://tinkerpop.apache.org/docs/current/reference/}{ Archive}
}
\usage{
data(Edges)
}
\description{
Two data frames, one for Vertex and Another with edges between those vertex from the example of Tinkerpop tutorial.
}
\examples{
data(Vertex)
content <- head(vertex)


}
\keyword{datasets}
