% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_parse_results_to_tables.R
\name{gremlin_parse_properties_to_data_frame}
\alias{gremlin_parse_properties_to_data_frame}
\title{Title}
\usage{
gremlin_parse_properties_to_data_frame(x)
}
\arguments{
\item{data}{}
}
\value{
a igraph object
}
\description{
Title
}
