% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_parse_results_to_tables.R
\name{gremlin_properties_to_table}
\alias{gremlin_properties_to_table}
\title{Pastes all properties of an element to a data frame}
\usage{
gremlin_properties_to_table(x)
}
\arguments{
\item{x}{Element}
}
\value{
Returns a data frame with all of the element's properties
}
\description{
Pastes all properties of an element to a data frame
}
