% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{package}
\name{neptune}
\alias{neptune}
\alias{neptune-package}
\title{Functions to Interact with a AWS Neptune Instance}
\description{
This package contains functions to  interact with AWS Neptune  - a graph
databse.
}
\section{Databases}{

}

\author{
Aleksander Dietrichson
}
