% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_structure_data.R
\name{neptune_count_elements}
\alias{neptune_count_elements}
\title{Counts total of edges and vertices}
\usage{
neptune_count_elements(ep, element = "both")
}
\arguments{
\item{ep}{Endpoint}

\item{element}{Select vertices, edges or both}
}
\value{
Displays ammount of vertices, edges or both
}
\description{
Counts total of edges and vertices
}
