% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_structure_data.R
\name{neptune_count_label_elements}
\alias{neptune_count_label_elements}
\title{Counts elements with specified label}
\usage{
neptune_count_label_elements(ep, element = "vertices", label)
}
\arguments{
\item{ep}{Endpoint}

\item{element}{Select vertices or edges}

\item{label}{Label to count}
}
\value{
Returns the number of elements with specified label
}
\description{
Counts elements with specified label
}
