% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_structure_data.R
\name{neptune_count_property}
\alias{neptune_count_property}
\title{Count number of elements with chosen value in property}
\usage{
neptune_count_property(ep, element = "Vertices", property, value)
}
\arguments{
\item{ep}{Endpoint}

\item{element}{Select vertices or edges}

\item{property}{Name of the property to search}

\item{value}{Specific value of selected property to search}
}
\value{
Returns number of results for selected property value.
}
\description{
Count number of elements with chosen value in property
}
\details{
Note: You need to use '' when passing property and value (for example: property="'name'",value="'marko'");
If the property is a reserved name (ID, for example), or the value is not a string, don't use '' (for example: 
property="id",value="'aaab111'" or property="'age'",value="40")
}
