% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_download.R
\name{neptune_download}
\alias{neptune_download}
\title{Full Download data from database}
\usage{
neptune_download(ep, element = "Vertices")
}
\arguments{
\item{ep}{Endpoint}

\item{element}{Element type (Vertices or Edges)}
}
\value{
Data frame with all selected elements
}
\description{
Full Download data from database
}
