% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_structure_data.R
\name{neptune_get_labels}
\alias{neptune_get_labels}
\title{Returns all unique labels for edges and vertices}
\usage{
neptune_get_labels(ep, element = "vertices")
}
\arguments{
\item{ep}{Endpoint}

\item{element}{Select vertices or edges}
}
\value{
Returns the unique labels of selected element
}
\description{
Returns all unique labels for edges and vertices
}
