% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_load_data.R
\name{neptune_load_data}
\alias{neptune_load_data}
\title{Load data into your Neptune Instance}
\usage{
neptune_load_data(data, endpoint, S3_endpoint, AWS_accessKey, AWS_secret)
}
\arguments{
\item{AWS_secret}{}
}
\description{
Load data into your Neptune Instance
}
