% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_load_data.R
\name{neptune_load_vertex}
\alias{neptune_load_vertex}
\title{Load data into your Neptune Instance from a data frame}
\usage{
neptune_load_vertex(data, ep, definedid = TRUE)
}
\arguments{
\item{data}{The source data frame}

\item{ep}{Neptune Endpoint}

\item{definedid}{Define if the ID is specified in the data (otherwise, it should be set automatically)}
}
\value{
exported data
}
\description{
Load data into your Neptune Instance from a data frame
}
