% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_parse_results_to_tables.R
\name{neptune_parse_results_to_tables}
\alias{neptune_parse_results_to_tables}
\title{Title}
\usage{
neptune_parse_results_to_tables(data)
}
\arguments{
\item{data}{}
}
\value{
a igraph object
}
\description{
Title
}
