% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neptune_download.R
\name{neptune_range_download}
\alias{neptune_range_download}
\title{Download data from database within a defined range}
\usage{
neptune_range_download(ep, from, to, element = "Vertices")
}
\arguments{
\item{ep}{Endpoint}

\item{from}{Inferior limit of the range}

\item{to}{Upper limit of the range}

\item{element}{Element type (Vertices or Edges)}
}
\value{
Data frame with all selected elements within the range
}
\description{
Download data from database within a defined range
}
