% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_json_query.R
\name{neptune_raw_json_query}
\alias{neptune_raw_json_query}
\title{Send a Raw query to Neptune}
\usage{
neptune_raw_json_query(ep, jsonQuery, as = c("parsed", "text", "raw"), ...)
}
\arguments{
\item{ep}{Endpoint}

\item{jsonQuery}{The Raw Query (in JSON) to be sent}

\item{as}{how to return the results}
}
\value{
the result from POST (in JSON)
}
\description{
Send a Raw query to Neptune
}
