% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_raw_query.R
\name{neptune_send_json_query}
\alias{neptune_send_json_query}
\title{Send a Raw query to Neptune}
\usage{
neptune_send_json_query(ep, query)
}
\arguments{
\item{ep}{Endpoint}

\item{query}{The Raw Query (in JSON) to be sent}
}
\value{
the result from POST (in JSON)
}
\description{
Send a Raw query to Neptune
}
