% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_neptune_endpoint.R
\name{print.neptune_endpoint}
\alias{print.neptune_endpoint}
\title{Print method for Neptune Endpoint}
\usage{
\method{print}{couch_connection}(ep)
}
\arguments{
\item{ep}{A neptune endpoint}
}
\description{
Prints theNeptune Endpoint object.
}
