% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_raw_query.R
\name{neptune_send_json_query}
\alias{neptune_send_json_query}
\title{Send a Query to Neptune}
\usage{
neptune_send_json_query(ep, query, as = c("parsed", "text", "raw"), ...)
}
\arguments{
\item{ep}{Endpoint}

\item{query}{The Query (in selected type) to be sent}
}
\value{
the result from POST (in JSON)
}
\description{
Send a Query to Neptune
}
